+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 # Unreleased
 
+- Add `ocaml.copy-type-under-cursor` to copy, in the clipboard the type of 
+  the expression under the cursor (#1582)
+
 ## 1.20.1
 
 - This was a version bump only, there were no changes